feat: faq section ad dag #692
