+++ src/routes/(public)/ad-dag/+page.svelte
@@ -4,7 +4,7 @@
 	// Import images
 	import { addag1, addag2, addag3 } from '$lib'
 
-	import { MultipleFaq, SingleFaq, Hero, Addag, Location, Schedule, RSectionHero, RPicture, RSectionPage, RSectionPlanning } from '$lib'
+	import { MultipleFaq, SingleFaq, Hero, Addag, Location, Schedule, RSectionHero, RPicture, RSectionPage, RSectionPlanning, RFaqSection } from '$lib'
 
 	// TODO: schedule data that needs to be added to directus
 	const scheduleData = {
@@ -33,6 +33,31 @@
 				startTime: '15:30',
 				endTime: '17:00',
 				event: 'Borrel & ontmoeting'
+			}
+		]
+	}
+
+	// TODO: FAQ that needs to be added to directus
+	const faqData = {
+		faqs: [
+			{
+				question: 'Voor wie is de Ad-dag bedoeld?',
+				answer:
+					'De Landelijke Ad-dag is bedoeld voor iedereen die betrokken is bij of interesse heeft in Associate degree-onderwijs in Nederland. Dit omvat studenten, docenten, onderwijsprofessionals, werkveldpartners, onderwijsinstellingen en andere geïnteresseerden die willen zien hoe Ad-onderwijs impact maakt op onderwijs en arbeidsmarkt.'
+			},
+			{
+				question: 'Wie komen er naar de Ad-dag?',
+				answer:
+					'De Landelijke Ad-dag brengt alle betrokkenen van Associate degree-onderwijs samen, zoals studenten en alumni, docenten en onderwijsprofessionals, werkveldpartners en werkgevers, en beleidsmakers en onderwijsorganisaties die kennis en ervaringen willen uitwisselen. Het is een dag vol inspiratie, netwerken en het delen van de impact van Ad-onderwijs.'
+			},
+			{
+				question: 'Waarom zou ik naar de Ad-dag moeten gaan?',
+				answer:
+					'De Landelijke Ad-dag is dé gelegenheid om inspiratie op te doen, te netwerken met studenten, docenten, werkveldpartners en beleidsmakers, en te ontdekken hoe Associate degree-onderwijs impact maakt op zowel onderwijs als arbeidsmarkt. Je leert over innovatie, doorstroommogelijkheden en krijgt inzichten die je direct kunt toepassen in je studie, werk of organisatie.'
+			},
+			{
+				question: 'Wanneer en waar vindt de Ad-dag plaats?',
+				answer: 'De Landelijke Ad-dag 2026 vindt plaats op vrijdag 17 april 2026 in Groningen, georganiseerd bij de Hanzehogeschool.'
 			}
 		]
 	}
@@ -72,36 +97,12 @@
 	}}
 />
 
-<section class="info-addag">
-	<img
-		class="logo"
-		src={logomobile}
-		alt=""
+<section class="faq-addag">
+	<RFaqSection
+		title="Alles wat je wilt weten over de Ad-dag?"
+		{faqData}
+		backgroundBlack
 	/>
-	<h2>Alles wat je wilt weten over de Ad-dag?</h2>
-
-	<MultipleFaq>
-		<SingleFaq
-			open={true}
-			question="Voor wie is de Ad-dag bedoeld?"
-			answer="De Landelijke Ad‑dag is bedoeld voor iedereen die betrokken is bij of interesse heeft in Associate degree‑onderwijs in Nederland. Dit omvat studenten, docenten, onderwijsprofessionals, werkveldpartners, onderwijsinstellingen en andere geïnteresseerden die willen zien hoe Ad‑onderwijs impact maakt op onderwijs en arbeidsmarkt."
-		/>
-
-		<SingleFaq
-			question="Wie komen er naar de Ad-dag?"
-			answer="De Landelijke Ad‑dag brengt alle betrokkenen van Associate degree‑onderwijs samen, zoals studenten en alumni, docenten en onderwijsprofessionals, werkveldpartners en werkgevers, en beleidsmakers en onderwijsorganisaties die kennis en ervaringen willen uitwisselen. Het is een dag vol inspiratie, netwerken en het delen van de impact van Ad‑onderwijs."
-		/>
-
-		<SingleFaq
-			question="Waarom zou ik naar de Ad-dag moeten gaan?"
-			answer="De Landelijke Ad‑dag is dé gelegenheid om inspiratie op te doen, te netwerken met studenten, docenten, werkveldpartners en beleidsmakers, en te ontdekken hoe Associate degree‑onderwijs impact maakt op zowel onderwijs als arbeidsmarkt. Je leert over innovatie, doorstroommogelijkheden en krijgt inzichten die je direct kunt toepassen in je studie, werk of organisatie."
-		/>
-
-		<SingleFaq
-			question="Wanneer en waar vindt de Ad-dag plaats?"
-			answer="De Landelijke Ad‑dag 2026 vindt plaats op vrijdag 17 april 2026 in Groningen, georganiseerd bij de Hanzehogeschool."
-		/>
-	</MultipleFaq>
 
 	<div class="addag2">
 		<RPicture
@@ -139,85 +140,27 @@
 <Location />
 
 <style>
-	.intro {
-		display: flex;
-		flex-direction: column;
-		padding: 3em 5%;
-		max-width: 1400px;
-		align-items: center;
-		gap: 4em;
-		align-self: center;
-		margin: auto;
-
-		img {
-			width: 100%;
-			border-radius: 1em;
-		}
-
-		article {
-			max-width: 600px;
-
-			p {
-				margin-top: 1em;
-			}
-		}
-	}
-
-	.addag3 {
-		overflow: hidden;
-		border-radius: 1em;
-		max-width: 600px;
-	}
-
-	@media (min-width: 768px) {
-		.intro {
-			flex-direction: row;
-			gap: 2em;
-			align-items: center;
-			padding: 5em 5%;
-			gap: 4em;
-
-			img {
-				width: 600px;
-			}
-		}
-	}
-
-	.info-addag {
+	.faq-addag {
 		display: flex;
 		flex-direction: column;
 		background-color: light-dark(var(--blue-100), hsl(210, 30%, 8%));
-		gap: 2em;
 		padding: 3em 5%;
-		box-sizing: border-box;
-		position: relative;
 		box-shadow: 0 0 0 100vmax var(--blue-100);
 		clip-path: inset(0 -100vmax);
+		overflow: hidden;
 
-		.logo {
-			margin-top: 2em;
-			align-self: center;
-		}
 
-		h2 {
-			text-align: center;
-			margin: auto;
-			max-width: 350px;
-		}
-	}
-
-	.addag2 {
-		margin-top: 5em;
-		align-self: center;
-		object-fit: cover;
-		width: 100%;
-		border-radius: 5em;
-	}
-
-	@media (min-width: 768px) {
 		.addag2 {
-			height: 400px;
-			width: 1000px;
+			margin-top: 5em;
+			align-self: center;
+			object-fit: cover;
+			width: 100%;
+			border-radius: 5em;
+
+			@media (min-width: 768px) {
+				height: 400px;
+				max-width: 1000px;
+			}
 		}
 	}
 </style>
